+++ Beam/database/migrations/2019_11_07_153448_add_indexes_to_article_views_snapshots.php
@@ -1,0 +1,18 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddIndexesToArticleViewsSnapshots extends Migration
+{
+    public function up()
+    {
+        Schema::table('article_views_snapshots', function(Blueprint $table) {
+            $table->index(['external_article_id']);
+            $table->index(['property_token']);
+            // implicit name for index is too long, therefore naming explicitly
+            $table->index(['time', 'derived_referer_medium', 'explicit_referer_medium'], 'time_referer_mediums');
+        });
+    }
+}