+++ Beam/resources/views/dashboard/public.blade.php
@@ -22,6 +22,11 @@
             components: {
                 DashboardRoot
             },
+            created: function() {
+                this.$store.commit('changeSettings', {
+                    newGraph: {{ json_encode(config('beam.pageview_graph_data_source') === 'snapshots') }}
+                })
+            },
             provide: function() {
                 return {
                     dashboardOptions: this.options